+++ paimon-core/src/main/java/org/apache/paimon/migrate/FileMetaUtils.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.migrate.utils;
+package org.apache.paimon.migrate;
 
 import org.apache.paimon.data.BinaryRow;
 import org.apache.paimon.data.BinaryRowWriter;
